requirements.txt
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
fastapi>=0.104.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
python-multipart>=0.0.6
python-telegram-bot==22.6
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn>=0.24.0
